pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "universal-doc-processing"
version = "0.1.0"
description = "5-step document pipeline"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic>=2",
    "sqlmodel",
    "httpx",
    "tenacity",
    "python-dotenv",
    "loguru",
    "rapidfuzz",
    "python-dateutil",
    "ruff",
    "pytest",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
